# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gabinete Digital Cícero João
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/demands:
    get:
      tags: [Admin]
      operationId: listDemands
      summary: List and filter all demands (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: neighborhoodId
          in: query
          required: false
          schema:
            type: integer
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Demands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Demand"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/demands/{id}/activities:
    post:
      tags: [Admin]
      operationId: createDemandActivity
      summary: Add an update / change status of a demand (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandActivityInput"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemandActivity"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/dashboard:
    get:
      tags: [Admin]
      operationId: getDashboardSummary
      summary: Admin dashboard summary (counts, recent activity)
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /agenda:
    get:
      tags: [Public]
      operationId: listAgenda
      summary: List public agenda events
      responses:
        "200":
          description: Agenda events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgendaEvent"
    post:
      tags: [Admin]
      operationId: createAgendaEvent
      summary: Create an agenda event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgendaEventInput"
      responses:
        "201":
          description: Agenda event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaEvent"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /agenda/{id}:
    patch:
      tags: [Admin]
      operationId: updateAgendaEvent
      summary: Update an agenda event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgendaEventUpdate"
      responses:
        "200":
          description: Agenda event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaEvent"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Admin]
      operationId: deleteAgendaEvent
      summary: Delete an agenda event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /appointments:
    post:
      tags: [Citizen]
      operationId: createAppointment
      summary: Request an in-person service appointment (atendimento)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /admin/integrations:
    get:
      tags: [Admin]
      operationId: listIntegrationStatus
      summary: Configuration status of external integrations (WhatsApp, ElevenLabs, HeyGen)
      responses:
        "200":
          description: Integration statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntegrationStatus"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/ai/voice:
    post:
      tags: [Admin]
      operationId: generateVoice
      summary: Synthesize speech via ElevenLabs (returns not-configured when no key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceRequest"
      responses:
        "200":
          description: Result (ok=false with error when not configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/ai/video:
    post:
      tags: [Admin]
      operationId: generateVideo
      summary: Generate an avatar video via HeyGen (returns not-configured when no key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoRequest"
      responses:
        "200":
          description: Result (ok=false with error when not configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /me/appointments:
    get:
      tags: [Citizen]
      operationId: listMyAppointments
      summary: List the current citizen's appointment requests
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/appointments:
    get:
      tags: [Admin]
      operationId: listAppointments
      summary: List all appointment requests (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/appointments/{id}:
    patch:
      tags: [Admin]
      operationId: updateAppointment
      summary: Update an appointment (confirm/schedule/cancel) (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Demand:
      type: object
      required:
        - id
        - protocol
        - title
        - description
        - status
        - citizenName
        - createdAt
      properties:
        id:
          type: integer
        protocol:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          description: "recebida | em_analise | encaminhada | aguardando_resposta | em_acompanhamento | resolvida | arquivada"
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        neighborhoodId:
          type: ["integer", "null"]
        neighborhoodName:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        citizenName:
          type: string
        citizenEmail:
          type: ["string", "null"]
        citizenPhone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    DemandActivity:
      type: object
      required: [id, demandId, note, createdAt]
      properties:
        id:
          type: integer
        demandId:
          type: integer
        status:
          type: ["string", "null"]
        note:
          type: string
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    DemandActivityInput:
      type: object
      required: [note]
      properties:
        status:
          type: string
        note:
          type: string
          minLength: 1
...
    AgendaEvent:
      type: object
      required: [id, title, type, startsAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: string
          description: "visita | reuniao | audiencia | sessao | outro"
        startsAt:
          type: string
          format: date-time
    AgendaEventInput:
      type: object
      required: [title, type, startsAt]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
        type:
          type: string
        startsAt:
          type: string
          format: date-time
    AgendaEventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        location:
          type: string
        type:
          type: string
        startsAt:
          type: string
          format: date-time
    Appointment:
      type: object
      required: [id, protocol, subject, status, citizenName, createdAt]
      properties:
        id:
          type: integer
        protocol:
          type: string
        subject:
          type: string
        description:
          type: ["string", "null"]
        preferredDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          description: "solicitado | confirmado | realizado | cancelado"
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        citizenName:
          type: string
        citizenEmail:
          type: ["string", "null"]
        citizenPhone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AppointmentInput:
      type: object
      required: [subject, citizenName]
      properties:
        subject:
          type: string
          minLength: 1
        description:
          type: string
        slotId:
          type: integer
        preferredDate:
          type: string
          format: date
        citizenName:
          type: string
          minLength: 1
        citizenEmail:
          type: string
        citizenPhone:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
        scheduledAt:
          type: string
          format: date-time
...
    IntegrationStatus:
      type: object
      required: [provider, configured, missing]
      properties:
        provider:
          type: string
        configured:
          type: boolean
        missing:
          type: array
          items:
            type: string
    VoiceRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
        voiceId:
          type: ["string", "null"]
    VoiceResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        audioBase64:
          type: ["string", "null"]
        contentType:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
    VideoRequest:
      type: object
      required: [script]
      properties:
        script:
          type: string
          minLength: 1
        avatarId:
          type: ["string", "null"]
    VideoResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        videoId:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
    DashboardSummary:
      type: object
      required:
        - totalDemands
        - resolvedDemands
        - openDemands
        - pendingAppointments
        - byStatus
        - recentDemands
      properties:
        totalDemands:
          type: integer
        resolvedDemands:
          type: integer
        openDemands:
          type: integer
        pendingAppointments:
          type: integer
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        recentDemands:
          type: array
          items:
            $ref: "#/components/schemas/Demand"
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
...